clean
* system table int->short, in case of sharding errors.
* abiType reference upgrade.

+++ WeBASE-Collect-Bee-core/src/main/java/com/webank/webasebee/core/api/controller/BlockTaskPoolController.java
@@ -52,7 +52,7 @@ public class BlockTaskPoolController {
     @RequestMapping("/blocks/get")
     @ApiOperation(value = "Get finished block count", httpMethod = "GET")
     public long getFinishedBlockCount() {
-        return blockTaskPoolRepository.countBySyncStatus(TxInfoStatusEnum.DONE.getStatus());
+        return blockTaskPoolRepository.countBySyncStatus((short) TxInfoStatusEnum.DONE.getStatus());
     }
 
     @ResponseBody
